requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
multidict==5.2.0
oauth2client==4.1.3
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.0
